requirements.txt
faiss-cpu
joblib
networkx
numpy
ollama
openai
pandas
psycopg2
psycopg2_binary
scikit_learn
torch
transformers
